+++ admin/diagnostic.php
@@ -36,7 +36,7 @@ $checks = array(
 
 llxHeader('', $langs->trans('Diagnostic'));
 $head = emergencyhouseAdminPrepareHead();
-print dol_get_fiche_head($head, 'diagnostic', $langs->trans('EmergencyHouseSetup'), -1, 'emergencyhouse@emergencyhouse');
+print dol_get_fiche_head($head, 'diagnostic', $langs->trans('EmergencyHouseSetup'), -1, 'fontawesome_house-user');
 print load_fiche_titre($langs->trans('Diagnostic'), emergencyhouseAdminLinkBack(), 'technic');
 
 print '<table class="noborder centpercent">';